+++ projects/schematics/src/ng-add/index.spec.ts
@@ -10,14 +10,13 @@ describe('ng-add', () => {
   it('adds v6 dependencies to package.json', async () => {
     const { runner, appTree } = await setup();
 
-    const tree = await runner.runSchematicAsync<Schema>('ng-add', {}, appTree).toPromise();
+    const tree = await runner.runSchematic<Schema>('ng-add', { project: 'test-app' }, appTree);
 
     const packageJson = JSON.parse(tree.readContent('package.json'));
     expect(packageJson.dependencies).toBeDefined();
 
     const dependencies = packageJson.dependencies;
 
-    expect(dependencies['@fortawesome/fontawesome-svg-core']).toBe(iconPackVersion);
     expect(dependencies['@fortawesome/free-solid-svg-icons']).toBe(iconPackVersion);
     expect(dependencies['@fortawesome/angular-fontawesome']).toBe(angularFontawesomeVersion);
   });
@@ -25,32 +24,39 @@ describe('ng-add', () => {
   it('adds v5 dependencies to package.json', async () => {
     const { runner, appTree } = await setup();
 
-    const tree = await runner.runSchematicAsync<Schema>('ng-add', { version: '5' }, appTree).toPromise();
+    const tree = await runner.runSchematic<Schema>('ng-add', { project: 'test-app', version: '5' }, appTree);
 
     const packageJson = JSON.parse(tree.readContent('package.json'));
     expect(packageJson.dependencies).toBeDefined();
 
     const dependencies = packageJson.dependencies;
 
-    expect(dependencies['@fortawesome/fontawesome-svg-core']).toBe(v5.svgCoreVersion);
     expect(dependencies['@fortawesome/free-solid-svg-icons']).toBe(v5.iconPackVersion);
     expect(dependencies['@fortawesome/angular-fontawesome']).toBe(angularFontawesomeVersion);
   });
 
-  it('adds FontAwesomeModule import to the AppModule', async () => {
-    const { runner, appTree } = await setup();
+  it('adds FontAwesomeModule import to the AppModule when standalone=false', async () => {
+    const { runner, appTree } = await setup(false);
 
-    const tree = await runner.runSchematicAsync<Schema>('ng-add', {}, appTree).toPromise();
+    const tree = await runner.runSchematic<Schema>('ng-add', { project: 'test-app' }, appTree);
 
     const contents = tree.readContent('src/app/app.module.ts');
 
     expect(contents).toContain('import { FontAwesomeModule }');
   });
 
+  it('does not attempt to add FontAwesomeModule import to the AppModule when standalone=true', async () => {
+    const { runner, appTree } = await setup();
+
+    const tree = await runner.runSchematic<Schema>('ng-add', { project: 'test-app' }, appTree);
+
+    expect(tree.files).not.toContain('/src/app/app.module.ts');
+  });
+
   it('installs @fortawesome/free-solid-svg-icons package by default', async () => {
     const { runner, appTree } = await setup();
 
-    const tree = await runner.runSchematicAsync<Schema>('ng-add', {}, appTree).toPromise();
+    const tree = await runner.runSchematic<Schema>('ng-add', { project: 'test-app' }, appTree);
 
     const packageJson = JSON.parse(tree.readContent('package.json'));
     expect(packageJson.dependencies).toBeDefined();
@@ -62,17 +68,14 @@ describe('ng-add', () => {
   it('allows to install several @fortawesome/*-svg-icons packages', async () => {
     const { runner, appTree } = await setup();
 
-    const tree = await runner
-      .runSchematicAsync<Schema>(
-        'ng-add',
-        {
-          project: 'test-app',
-          iconPackages: ['free-solid', 'free-brands', 'free-regular'],
-        },
-        appTree,
-      )
-      .toPromise();
-
+    const tree = await runner.runSchematic<Schema>(
+      'ng-add',
+      {
+        project: 'test-app',
+        iconPackages: ['free-solid', 'free-brands', 'free-regular'],
+      },
+      appTree,
+    );
     const packageJson = JSON.parse(tree.readContent('package.json'));
     expect(packageJson.dependencies).toBeDefined();
 
@@ -85,17 +88,14 @@ describe('ng-add', () => {
   it('allows to install no icon packages', async () => {
     const { runner, appTree } = await setup();
 
-    const tree = await runner
-      .runSchematicAsync<Schema>(
-        'ng-add',
-        {
-          project: 'test-app',
-          iconPackages: [],
-        },
-        appTree,
-      )
-      .toPromise();
-
+    const tree = await runner.runSchematic<Schema>(
+      'ng-add',
+      {
+        project: 'test-app',
+        iconPackages: [],
+      },
+      appTree,
+    );
     const packageJson = JSON.parse(tree.readContent('package.json'));
     expect(packageJson.dependencies).toBeDefined();
 
@@ -104,21 +104,20 @@ describe('ng-add', () => {
   });
 });
 
-const setup = async () => {
+const setup = async (standalone?: boolean) => {
   const runner = new SchematicTestRunner('schematics', collectionPath);
 
-  const appTree = await runner
-    .runExternalSchematicAsync(
-      '@schematics/angular',
-      'ng-new',
-      {
-        name: 'test-app',
-        version: '9.0.0-rc.6',
-        directory: '.',
-      },
-      Tree.empty(),
-    )
-    .toPromise();
+  const appTree = await runner.runExternalSchematic(
+    '@schematics/angular',
+    'ng-new',
+    {
+      name: 'test-app',
+      version: '9.0.0-rc.6',
+      directory: '.',
+      standalone,
+    },
+    Tree.empty(),
+  );
 
   return { runner, appTree };
 };