+++ projects/schematics/src/ng-add/index.ts
@@ -8,7 +8,8 @@ import {
 import { addImportToModule } from '@schematics/angular/utility/ast-utils';
 import { InsertChange } from '@schematics/angular/utility/change';
 import { addPackageJsonDependency, NodeDependencyType } from '@schematics/angular/utility/dependencies';
-import { getAppModulePath } from '@schematics/angular/utility/ng-ast-utils';
+import { getAppModulePath, isStandaloneApp } from '@schematics/angular/utility/ng-ast-utils';
+import { getMainFilePath } from '@schematics/angular/utility/standalone/util';
 import { getWorkspace } from '@schematics/angular/utility/workspace';
 import { Schema } from './schema';
 import { angularFontawesomeVersion, iconPackVersion, v5 } from './versions';
@@ -16,12 +17,6 @@ import { angularFontawesomeVersion, iconPackVersion, v5 } from './versions';
 export default function (options: Schema): Rule {
   return chain([
     (tree: Tree, context: SchematicContext) => {
-      addPackageJsonDependency(tree, {
-        type: NodeDependencyType.Default,
-        name: '@fortawesome/fontawesome-svg-core',
-        version: options.version === '6' ? iconPackVersion : v5.svgCoreVersion,
-      });
-
       addPackageJsonDependency(tree, {
         type: NodeDependencyType.Default,
         name: '@fortawesome/angular-fontawesome',
@@ -48,27 +43,29 @@ export default function (options: Schema): Rule {
 function addModule(options: Schema): Rule {
   return async (host: Tree) => {
     const workspace = await getWorkspace(host);
-    const projectName = options.project ?? (workspace.extensions.defaultProject as string);
+    const projectName = options.project as string;
     const project = workspace.projects.get(projectName);
     if (project == null) {
       throw new SchematicsException(`Project with name ${projectName} does not exist.`);
     }
-    const buildOptions = getProjectTargetOptions(project, 'build');
-    const modulePath = getAppModulePath(host, buildOptions.main as string);
-    const moduleSource = getSourceFile(host, modulePath);
-    const changes = addImportToModule(
-      moduleSource,
-      modulePath,
-      'FontAwesomeModule',
-      '@fortawesome/angular-fontawesome',
-    );
-    const recorder = host.beginUpdate(modulePath);
-    changes.forEach((change) => {
-      if (change instanceof InsertChange) {
-        recorder.insertLeft(change.pos, change.toAdd);
-      }
-    });
-    host.commitUpdate(recorder);
+    const mainPath = await getMainFilePath(host, projectName);
+    if (!isStandaloneApp(host, mainPath)) {
+      const modulePath = getAppModulePath(host, mainPath);
+      const moduleSource = getSourceFile(host, modulePath);
+      const changes = addImportToModule(
+        moduleSource,
+        modulePath,
+        'FontAwesomeModule',
+        '@fortawesome/angular-fontawesome',
+      );
+      const recorder = host.beginUpdate(modulePath);
+      changes.forEach((change) => {
+        if (change instanceof InsertChange) {
+          recorder.insertLeft(change.pos, change.toAdd);
+        }
+      });
+      host.commitUpdate(recorder);
+    }
   };
 }