+++ projects/schematics/src/ng-add/versions.ts
@@ -1,7 +1,6 @@
-export const angularFontawesomeVersion = '~0.10.2';
-export const iconPackVersion = '^6.1.0';
+export const angularFontawesomeVersion = '~0.15.0';
+export const iconPackVersion = '^6.5.2';
 
 export const v5 = {
-  svgCoreVersion: '~1.2.36',
   iconPackVersion: '^5.15.4',
 };