+++ src/lib/layers/layers-counter.component.ts
@@ -1,11 +1,15 @@
-import { Component, HostBinding, Input, OnChanges, Optional, SimpleChanges } from '@angular/core';
+import { DOCUMENT } from '@angular/common';
+import { Component, HostBinding, inject, Input, OnChanges, Optional, SimpleChanges } from '@angular/core';
 import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
-import { counter, CounterParams, Styles } from '@fortawesome/fontawesome-svg-core';
+import { counter, CounterParams } from '@fortawesome/fontawesome-svg-core';
+import { FaConfig } from '../config';
 import { faWarnIfParentNotExist } from '../shared/errors/warn-if-parent-not-exist';
+import { ensureCss } from '../shared/utils/css';
 import { FaLayersComponent } from './layers.component';
 
 @Component({
   selector: 'fa-layers-counter',
+  standalone: true,
   template: '',
   host: {
     class: 'ng-fa-layers-counter',
@@ -14,13 +18,17 @@ import { FaLayersComponent } from './layers.component';
 export class FaLayersCounterComponent implements OnChanges {
   @Input() content: string;
   @Input() title?: string;
-  @Input() styles?: Styles;
-  @Input() classes?: string[] = [];
   @Input() position?: 'bottom-right' | 'bottom-left' | 'top-right' | 'top-left';
 
   @HostBinding('innerHTML') renderedHTML: SafeHtml;
 
-  constructor(@Optional() private parent: FaLayersComponent, private sanitizer: DomSanitizer) {
+  private document = inject(DOCUMENT);
+  private config = inject(FaConfig);
+
+  constructor(
+    @Optional() private parent: FaLayersComponent,
+    private sanitizer: DomSanitizer,
+  ) {
     faWarnIfParentNotExist(this.parent, 'FaLayersComponent', this.constructor.name);
   }
 
@@ -32,21 +40,14 @@ export class FaLayersCounterComponent implements OnChanges {
   }
 
   protected buildParams(): CounterParams {
-    const classes = [];
-    if (this.classes != null) {
-      classes.push(...this.classes);
-    }
-    if (this.position != null) {
-      classes.push(`fa-layers-${this.position}`);
-    }
     return {
       title: this.title,
-      classes,
-      styles: this.styles,
+      classes: this.position != null ? [`fa-layers-${this.position}`] : undefined,
     };
   }
 
   private updateContent(params: CounterParams) {
+    ensureCss(this.document, this.config);
     this.renderedHTML = this.sanitizer.bypassSecurityTrustHtml(counter(this.content || '', params).html.join(''));
   }
 }