+++ src/lib/layers/layers.component.ts
@@ -1,12 +1,25 @@
-import { Component, ElementRef, HostBinding, Input, OnChanges, OnInit, Renderer2, SimpleChanges } from '@angular/core';
+import { DOCUMENT } from '@angular/common';
+import {
+  Component,
+  ElementRef,
+  HostBinding,
+  inject,
+  Input,
+  OnChanges,
+  OnInit,
+  Renderer2,
+  SimpleChanges,
+} from '@angular/core';
 import { SizeProp } from '@fortawesome/fontawesome-svg-core';
 import { FaConfig } from '../config';
+import { ensureCss } from '../shared/utils/css';
 
 /**
  * Fontawesome layers.
  */
 @Component({
   selector: 'fa-layers',
+  standalone: true,
   template: `<ng-content></ng-content>`,
 })
 export class FaLayersComponent implements OnInit, OnChanges {
@@ -14,10 +27,17 @@ export class FaLayersComponent implements OnInit, OnChanges {
 
   @Input() @HostBinding('class.fa-fw') fixedWidth?: boolean;
 
-  constructor(private renderer: Renderer2, private elementRef: ElementRef, private config: FaConfig) {}
+  private document = inject(DOCUMENT);
+
+  constructor(
+    private renderer: Renderer2,
+    private elementRef: ElementRef,
+    private config: FaConfig,
+  ) {}
 
   ngOnInit() {
     this.renderer.addClass(this.elementRef.nativeElement, 'fa-layers');
+    ensureCss(this.document, this.config);
     this.fixedWidth = typeof this.fixedWidth === 'boolean' ? this.fixedWidth : this.config.fixedWidth;
   }