+++ src/lib/public_api.ts
@@ -1,5 +1,5 @@
 export { FontAwesomeModule } from './fontawesome.module';
-export { FaProps } from './shared/models/props.model';
+export { AnimationProp, FaProps } from './shared/models/props.model';
 export { FaIconComponent } from './icon/icon.component';
 export { FaDuotoneIconComponent } from './icon/duotone-icon.component';
 export { FaConfig } from './config';
@@ -9,3 +9,16 @@ export { FaLayersCounterComponent } from './layers/layers-counter.component';
 export { FaStackComponent } from './stack/stack.component';
 export { FaStackItemSizeDirective } from './stack/stack-item-size.directive';
 export { FaIconLibrary, FaIconLibraryInterface } from './icon-library';
+export { IconPrefix, IconName, IconLookup, IconDefinition, IconPack } from './types';
+
+export {
+  IconParams,
+  CounterParams,
+  TextParams,
+  FaSymbol,
+  FlipProp,
+  PullProp,
+  RotateProp,
+  SizeProp,
+  Transform,
+} from '@fortawesome/fontawesome-svg-core';