+++ src/lib/shared/models/props.model.ts
@@ -1,22 +1,11 @@
-import {
-  FaSymbol,
-  FlipProp,
-  IconProp,
-  PullProp,
-  RotateProp,
-  SizeProp,
-  Styles,
-  Transform,
-} from '@fortawesome/fontawesome-svg-core';
+import { FlipProp, PullProp, RotateProp, SizeProp } from '@fortawesome/fontawesome-svg-core';
 
 /**
  * Fontawesome props.
  */
 export interface FaProps {
-  mask?: IconProp;
   className?: string;
-  spin?: boolean;
-  pulse?: boolean;
+  animation?: AnimationProp;
   border?: boolean;
   fixedWidth?: boolean;
   counter?: boolean;
@@ -25,8 +14,17 @@ export interface FaProps {
   size?: SizeProp;
   pull?: PullProp;
   rotate?: RotateProp;
-  transform?: string | Transform;
-  symbol?: FaSymbol;
-  style?: Styles;
   stackItemSize?: '1x' | '2x';
 }
+
+export type AnimationProp =
+  | 'beat'
+  | 'fade'
+  | 'beat-fade'
+  | 'bounce'
+  | 'flip'
+  | 'shake'
+  | 'spin'
+  | 'spin-reverse'
+  | 'spin-pulse'
+  | 'spin-pulse-reverse';