+++ src/testing/helpers.ts
@@ -1,7 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { Type } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { IconDefinition, IconName } from '@fortawesome/fontawesome-svg-core';
 import { FaDuotoneIconComponent } from '../lib/icon/duotone-icon.component';
 import { FaIconComponent } from '../lib/icon/icon.component';
 import { FaLayersCounterComponent } from '../lib/layers/layers-counter.component';
@@ -9,11 +8,12 @@ import { FaLayersTextComponent } from '../lib/layers/layers-text.component';
 import { FaLayersComponent } from '../lib/layers/layers.component';
 import { FaStackItemSizeDirective } from '../lib/stack/stack-item-size.directive';
 import { FaStackComponent } from '../lib/stack/stack.component';
+import { IconDefinition } from '../lib/types';
 
 export const initTest = <T>(component: Type<T>, providers?: any[]): ComponentFixture<T> => {
   TestBed.configureTestingModule({
-    imports: [CommonModule],
-    declarations: [
+    imports: [
+      CommonModule,
       FaIconComponent,
       FaDuotoneIconComponent,
       FaLayersComponent,
@@ -21,8 +21,8 @@ export const initTest = <T>(component: Type<T>, providers?: any[]): ComponentFix
       FaLayersCounterComponent,
       FaStackComponent,
       FaStackItemSizeDirective,
-      component,
     ],
+    declarations: [component],
     providers,
   });
   return TestBed.createComponent(component);
@@ -33,6 +33,6 @@ export const queryByCss = (fixture: ComponentFixture<any>, cssSelector: string):
 
 export const faDummy: IconDefinition = {
   prefix: 'fad',
-  iconName: 'dummy' as IconName,
-  icon: [512, 512, [], 'f030', ['M50 50 H412 V250 H50 Z', 'M50 262 H412 V462 H50 Z']],
+  iconName: 'dummy',
+  icon: [512, 512, [], '', ['M50 50 H412 V250 H50 Z', 'M50 262 H412 V462 H50 Z']],
 };